add separate widget for wordpair text widget.
The line responsible for showing the current
wordpair looks like this now:
Text(appState.current.asLowerCase).

To change it into something more complex,
it is a good idea to extract this line into a
separate widget. Having separate widgets for
separate logical parts of your UI is an important
way of managing complexity in Flutter.

Flutter provides a refactoring helper for
extracting widgets. Right click the piece of code
you want to refactor (Text in this case) and
select Refactor... from the drop-down menu,
In the Refactor menu, select Extract Widget.
Assign a name, such as BigCard, and click Enter.