Improve accessibility.
Flutter makes apps accessible by default.
For example, every Flutter app correctly surfaces
all text and interactive elements in the app to
screen readers such as TalkBack and VoiceOver.

Sometimes, though, some work is required. In the
case of this app the screen reader might have
problems pronouncing some generated word pairs.
humans don't have problems identifying the two
words in cheaphead, a screen reader might
pronounce the ph in the middle of the word as f.

A simple solution is to replace pair.asLowerCase
with "${pair.first} ${pair.second}". But it change
will change the UI and we want to do that.

so use Text's semanticsLabel property to override
the visual content of the text widget with a
semantic content that is more appropriate for
screen readers.